Reorganize configuration to be more flexible in the future
