+++ src/main/java/de/chojo/repbot/config/ConfigFile.java
@@ -1,6 +1,7 @@
 package de.chojo.repbot.config;
 
 import de.chojo.repbot.config.elements.Badges;
+import de.chojo.repbot.config.elements.BaseSettings;
 import de.chojo.repbot.config.elements.Botlist;
 import de.chojo.repbot.config.elements.Database;
 import de.chojo.repbot.config.elements.Links;
@@ -9,30 +10,20 @@ import de.chojo.repbot.config.elements.TestMode;
 
 @SuppressWarnings("FieldMayBeFinal")
 public class ConfigFile {
-    private String token = "";
-    private String defaultPrefix = "!";
+    private BaseSettings baseSettings = new BaseSettings();
     private Database database = new Database();
-    private boolean exclusiveHelp = false;
     private MagicImage magicImage = new MagicImage();
     private TestMode testMode = new TestMode();
     private Badges badges = new Badges();
     private Links links = new Links();
     private Botlist botlist = new Botlist();
 
-    public String token() {
-        return token;
-    }
-
-    public String defaultPrefix() {
-        return defaultPrefix;
+    public BaseSettings baseSettings() {
+        return baseSettings;
     }
 
     public Database database() {
         return database;
-    }
-
-    public boolean isExclusiveHelp() {
-        return exclusiveHelp;
     }
 
     public MagicImage magicImage() {
